pyproject.toml
[project]
name = "podcast-ads"
version = "0.1.0"
description = "AI-powered podcast ad stripper and transcriber"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai",
    "ffmpeg-python",
    "python-dotenv",
    "rich",
    "typer",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/podcast_ads"]
